src/requirements/base.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.3
django-admin-bootstrapped==2.5.7
django-authtools==1.6.0
django-braces==1.11.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.3
django-filter==1.0.4
django-filters==0.2.1
django-rest-swagger==2.1.2
django-treebeard==4.1.2
djangorestframework==3.6.3
easy-thumbnails==2.4.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
Pillow==4.2.1
prettyconf==1.2.3
pytz==2017.2
requests==2.18.1
simplejson==3.11.1
six==1.10.0
sqlparse==0.2.3
uritemplate==3.0.0
urllib3==1.21.1
Werkzeug==0.12.2
whitenoise==3.3.0
